Add offers.js, complete event.js
